fix(web): 🐛 Harden visual story recovery
